Updated profile route
Updated the app.route for the profile.html page to read the username of the current logged in user, query the user database and return the username and email address values. Cleaned these values to remove excess symbols so text would display as plain text in the html file, passed the values to the html page.