Feature front update profile (#101)
* added comment home.html line 83

* Updated profile route

Updated the app.route for the profile.html page to read the username of the current logged in user, query the user database and return the username and email address values. Cleaned these values to remove excess symbols so text would display as plain text in the html file, passed the values to the html page.

* Fix form errors and show user details

Removed all references to forms as this was all data linked to the old aws db. Added in a new section to just show the username and email address of the current logged in user.

---------

Co-authored-by: ShaneD <shane.dzartov@bigpond.com>