Redesigned the News page
I, Gimsara Kenula Elgiriyage, redesigned the news page according to the mockup I was given. I revised the page according to the feedback I got from Aditya.

Made the whole page responsive as I could.
Removed the background image.
Could not get the article routes working, must be an issue with the backend.
The navbar routes are working.
Made the current month filter to stand out.